Refactor: extract shared attribute matching utilities and remove unused code.
