+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/variation-selector/frontend.ts
@@ -21,6 +21,12 @@ import type {
 	AddToCartWithOptionsStore,
 	Context as AddToCartWithOptionsStoreContext,
 } from '../frontend';
+import {
+	normalizeAttributeName,
+	attributeNamesMatch,
+	getVariationAttributeValue,
+	findMatchingVariation,
+} from '../../../base/utils/variations/attribute-matching';
 import setStyles from './set-styles';
 
 type Option = {
@@ -49,57 +55,6 @@ const { state: productDataState } = store< ProductDataStore >(
 	{},
 	{ lock: universalLock }
 );
-
-/**
- * Normalize attribute name by stripping the 'attribute_' or 'attribute_pa_' prefix
- * that WooCommerce adds for variation attributes.
- *
- * @param name The attribute name (e.g., 'attribute_color' or 'attribute_pa_color').
- * @return The normalized name (e.g., 'color').
- */
-const normalizeAttributeName = ( name: string ): string => {
-	return name.replace( /^attribute_(pa_)?/, '' );
-};
-
-/**
- * Get the attribute value from a variation's attributes array.
- *
- * The Store API returns the attribute label (e.g., "Color") in the name field,
- * while the PHP context uses the attribute slug (e.g., "attribute_pa_color").
- * We do a case-insensitive comparison to match "color" with "Color".
- *
- * @param variation     The variation in Store API format.
- * @param attributeName The attribute name to find (may include 'attribute_' prefix).
- * @return The attribute value, or undefined if not found.
- */
-const getVariationAttributeValue = (
-	variation: ProductResponseItem[ 'variations' ][ number ],
-	attributeName: string
-): string | undefined => {
-	const normalizedName =
-		normalizeAttributeName( attributeName ).toLowerCase();
-	const attr = variation.attributes.find(
-		( a ) => a.name.toLowerCase() === normalizedName
-	);
-	return attr?.value;
-};
-
-/**
- * Check if two attribute names match, using case-insensitive comparison.
- *
- * This handles the mismatch between Store API labels (e.g., "Color") and
- * PHP context slugs (e.g., "attribute_pa_color").
- *
- * @param name1 First attribute name (may be label or slug format).
- * @param name2 Second attribute name (may be label or slug format).
- * @return True if the names match after normalization.
- */
-const attributeNamesMatch = ( name1: string, name2: string ): boolean => {
-	return (
-		normalizeAttributeName( name1 ).toLowerCase() ===
-		normalizeAttributeName( name2 ).toLowerCase()
-	);
-};
 
 /**
  * Check if the attribute value is valid given the other selected attributes and
@@ -454,30 +409,9 @@ const { actions, state } = store< VariableProductAddToCartWithOptionsStore >(
 				}
 
 				const { selectedAttributes } = getContext< Context >();
-
-				// Find matching variation ID from Store API format.
-				// Use case-insensitive comparison since Store API returns labels (e.g., "Color")
-				// while PHP context uses slugs (e.g., "attribute_pa_color" → "color").
-				const matchedVariation = product.variations.find(
-					( variation ) => {
-						return variation.attributes.every( ( attr ) => {
-							const attrNameLower = attr.name.toLowerCase();
-							const selectedAttr = selectedAttributes.find(
-								( selected ) =>
-									normalizeAttributeName(
-										selected.attribute
-									).toLowerCase() === attrNameLower
-							);
-							// Empty value means "Any" - matches any selection.
-							if ( attr.value === '' ) {
-								return (
-									selectedAttr !== undefined &&
-									selectedAttr.value !== ''
-								);
-							}
-							return selectedAttr?.value === attr.value;
-						} );
-					}
+				const matchedVariation = findMatchingVariation(
+					product,
+					selectedAttributes
 				);
 
 				const { actions: productDataActions } =
@@ -501,29 +435,9 @@ const { actions, state } = store< VariableProductAddToCartWithOptionsStore >(
 				}
 
 				const { selectedAttributes } = getContext< Context >();
-
-				// Find matching variation from Store API format.
-				// Use case-insensitive comparison since Store API returns labels (e.g., "Color")
-				// while PHP context uses slugs (e.g., "attribute_pa_color" → "color").
-				const matchedVariation = product.variations.find(
-					( variation ) => {
-						return variation.attributes.every( ( attr ) => {
-							const attrNameLower = attr.name.toLowerCase();
-							const selectedAttr = selectedAttributes.find(
-								( selected ) =>
-									normalizeAttributeName(
-										selected.attribute
-									).toLowerCase() === attrNameLower
-							);
-							if ( attr.value === '' ) {
-								return (
-									selectedAttr !== undefined &&
-									selectedAttr.value !== ''
-								);
-							}
-							return selectedAttr?.value === attr.value;
-						} );
-					}
+				const matchedVariation = findMatchingVariation(
+					product,
+					selectedAttributes
 				);
 
 				const { errorMessages } = getConfig();
